server/requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.29.0
yt-dlp==2024.5.27
spotdl==4.2.5
mutagen==1.47.0
aiofiles==23.2.1
python-multipart==0.0.9
Pillow==10.3.0
httpx==0.27.0
aiosqlite==0.20.0
SQLAlchemy[asyncio]==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0